refactor: Refactored Card Component - Moved Avatar into its own directory in src/components - Created a Layout components directory
