+++ src/components/Card/Card.tsx
@@ -21,14 +21,3 @@ export const CardImage = ({ src, alt, className }: CardImageProps) => {
 
   return <img src={src} alt={alt} className={classes} />;
 };
-interface AvatarProps {
-  src: string;
-  alt: string;
-  className?: string;
-}
-
-export const Avatar = ({ src, alt, className }: AvatarProps) => {
-  const classes = ["", className].join(" ").trim();
-
-  return <img src={src} alt={alt} className={classes} />;
-};
